requirements.txt
pandas>=2.0.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
openpyxl>=3.1.0
xlrd==1.2.0

